More stuff.
Still more to do but i was busy fixing my gmod addon and being sick.

most of this is just general stuff for the update but i did redo how hearts work so now they are properly replaced with werewolf or merfolk hears.

i was going to only commit this once everything was done but i have to fix an issue with 1.20.x with the devourer body entity.